+++ src/encrypt/mod.rs
@@ -21,7 +21,7 @@ use crate::{
     cbor::value::Value,
     iana,
     util::{cbor_type_error, AsCborValue},
-    CoseError, Header, ProtectedHeader,
+    CoseError, Header, ProtectedHeader, Result,
 };
 use alloc::{borrow::ToOwned, vec, vec::Vec};
 
@@ -48,7 +48,7 @@ pub struct CoseRecipient {
 impl crate::CborSerializable for CoseRecipient {}
 
 impl AsCborValue for CoseRecipient {
-    fn from_cbor_value(value: Value) -> Result<Self, CoseError> {
+    fn from_cbor_value(value: Value) -> Result<Self> {
         let mut a = match value {
             Value::Array(a) => a,
             v => return cbor_type_error(&v, "array"),
@@ -85,7 +85,7 @@ impl AsCborValue for CoseRecipient {
         })
     }
 
-    fn to_cbor_value(self) -> Result<Value, CoseError> {
+    fn to_cbor_value(self) -> Result<Value> {
         let mut v = vec![
             self.protected.cbor_bstr()?,
             self.unprotected.to_cbor_value()?,
@@ -236,7 +236,7 @@ impl crate::TaggedCborSerializable for CoseEncrypt {
 }
 
 impl AsCborValue for CoseEncrypt {
-    fn from_cbor_value(value: Value) -> Result<Self, CoseError> {
+    fn from_cbor_value(value: Value) -> Result<Self> {
         let mut a = match value {
             Value::Array(a) => a,
             v => return cbor_type_error(&v, "array"),
@@ -267,7 +267,7 @@ impl AsCborValue for CoseEncrypt {
         })
     }
 
-    fn to_cbor_value(self) -> Result<Value, CoseError> {
+    fn to_cbor_value(self) -> Result<Value> {
         let mut arr = Vec::new();
         for r in self.recipients {
             arr.push(r.to_cbor_value()?);
@@ -381,7 +381,7 @@ impl crate::TaggedCborSerializable for CoseEncrypt0 {
 }
 
 impl AsCborValue for CoseEncrypt0 {
-    fn from_cbor_value(value: Value) -> Result<Self, CoseError> {
+    fn from_cbor_value(value: Value) -> Result<Self> {
         let mut a = match value {
             Value::Array(a) => a,
             v => return cbor_type_error(&v, "array"),
@@ -403,7 +403,7 @@ impl AsCborValue for CoseEncrypt0 {
         })
     }
 
-    fn to_cbor_value(self) -> Result<Value, CoseError> {
+    fn to_cbor_value(self) -> Result<Value> {
         Ok(Value::Array(vec![
             self.protected.cbor_bstr()?,
             self.unprotected.to_cbor_value()?,